Changes made in DELL

+++ FrunutStock.Data/Infrastructure/RepositoryBase.cs
@@ -56,7 +56,7 @@ namespace FrunutStock.Data.Infrastructure
                 dbSet.Remove(obj);
         }
 
-        public virtual T GetById(int id)
+        public virtual T GetById(Int64 id)
         {
             return dbSet.Find(id);
         }
